基于日程结束时间是否过期，判断是否发送提醒通知

+++ interfaces/inner_api/reminder_request_calendar.h
@@ -147,7 +147,7 @@ public:
     virtual bool HandleSysTimeChange(uint64_t oriTriggerTime) override;
     static const uint8_t MAX_MONTHS_OF_YEAR;
     static const uint8_t MAX_DAYS_OF_MONTH;
-    static const uint8_t DELAY_REMINDER;
+    static const uint32_t DELAY_REMINDER;
     static void AppendValuesBucket(const sptr<ReminderRequest> &reminder,
         const sptr<NotificationBundleOption> &bundleOption, NativeRdb::ValuesBucket &values);
     static uint8_t GetDaysOfMonth(const uint16_t &year, const uint8_t &month);
